+++ frontend/app/events/event/event.component.scss
@@ -1,0 +1,4 @@
+.event-card {
+  overflow: scroll;
+  width: 100%;
+}
