+++ frontend/app/events/event/index.ts
@@ -1,0 +1,1 @@
+export { EventComponent } from "./event.component"
