+++ frontend/app/events/eventDetail/eventDetail.component.html
