+++ frontend/app/events/eventDetail/eventDetail.component.scss
