+++ frontend/app/events/eventDetail/index.ts
@@ -1,0 +1,1 @@
+export { EventDetailComponent } from "./eventDetail.component"
