+++ frontend/app/events/index.ts
@@ -1,4 +1,3 @@
-export { Event } from "./event"
-export { EventService } from "./event.service"
-export { EventComponent } from "./event.component"
-export { EventDetailComponent } from "./eventDetail.component"
+export * from "./shared"
+export * from "./event"
+export * from "./eventDetail"
