+++ frontend/app/events/shared/event.service.ts
