+++ frontend/app/events/shared/index.ts
@@ -1,0 +1,2 @@
+export { Event } from "./event"
+export { EventService } from "./event.service"
