+++ gulp.config.js
@@ -98,6 +98,9 @@ module.exports = function () {
           'app/captcha': { main: 'index.js' },
           'app/contact': { main: 'index.js' },
           'app/events': { main: 'index.js' },
+          'app/events/shared': { main: 'index.js' },
+          'app/events/eventDetail': { main: 'index.js' },
+          'app/events/event': { main: 'index.js' },
           'app/facebook': { main: 'index.js' },
           'app/map': { main: 'index.js'},
           'app/imprint': { main: 'index.js'},
