+++ systemjs.config.js
@@ -23,6 +23,9 @@
     'app/captcha',
     'app/contact',
     'app/events',
+    'app/events/shared',
+    'app/events/eventDetail',
+    'app/events/event',
     'app/facebook',
     'app/map',
     'app/imprint',
